Added DisplayOrder reordering
Added DisplayOrder reordering for when you edit a Product, and when you
delete a Product, it fixes the DisplayOrder so that it is consecutive.
Added a repository layer and services to do this, and I will probably
refactor the code later so that the controller uses the repository layer
and services for everything, to keep the controller code clean.
Also added input validation for the Product Unit